Move JSON-LD cache helpers to runtime
Keep generated vocabulary classes focused on schema-specific constants by
importing reusable JSON-LD cache and IRI trust helpers from
@fedify/vocab-runtime.  The generated code now only emits the portable IRI key
set and pattern, then passes them to the runtime normalization helper.

This also gives the cache compaction and IRI normalization paths direct unit
coverage outside generated snapshots.

https://github.com/fedify-dev/fedify/pull/850#pullrequestreview-4616181669
